Modul 21
The ability for users to subscribe and unsubscribe to the newsletter of any category. Notification of subscription/unsubscription. Notification of a new post or post update. Every week, the user receives an email with a list of new posts that appeared during the week. a welcome letter to the user upon registration in the application. Email confirmation during registration. The ability to create only 3 publications per day.